backend/requirements.txt
drf-extra-fields==3.2.1
django-colorfield==0.10.1
django-filter==21.1
djangorestframework==3.14.0
djoser==2.1.0
gunicorn==20.1.0
python-decouple==3.5
Pillow==9.3.0
psycopg2-binary==2.9.3
reportlab==4.0.7
